Move _unpack_models to widget.js and widget serialization to widget.py
